+++ joblib/parallel.py
@@ -165,13 +165,13 @@ class parallel_backend(object):
     threads usable in some C-level library threadpools like openBLAS, MKL or
     OpenMP. The default limit in each worker is set to
     ``cpu_count() // effective_n_jobs`` but this limit can be overwritten with
-    the ``max_inner_num_threads`` argument which will be used to set this limit
+    the ``inner_max_num_threads`` argument which will be used to set this limit
     in the child processes.
 
     .. versionadded:: 0.10
 
     """
-    def __init__(self, backend, n_jobs=-1, max_inner_num_threads=None,
+    def __init__(self, backend, n_jobs=-1, inner_max_num_threads=None,
                  **backend_params):
         if isinstance(backend, _basestring):
             if backend not in BACKENDS and backend in EXTERNAL_BACKENDS:
@@ -180,11 +180,11 @@ class parallel_backend(object):
 
             backend = BACKENDS[backend](**backend_params)
 
-        if max_inner_num_threads is not None:
-            msg = ("max_inner_num_threads argument should only be used with "
+        if inner_max_num_threads is not None:
+            msg = ("inner_max_num_threads argument should only be used with "
                    "backend='loky'")
-            assert backend.support_inner_num_threads, msg
-            backend.max_inner_num_threads = max_inner_num_threads
+            assert backend.support_inner_max_num_threads, msg
+            backend.inner_max_num_threads = inner_max_num_threads
 
         # If the nesting_level of the backend is not set previously, use the
         # nesting level from the previous active_backend to set it